requirements.txt
google-cloud-storage
gsutil
numpy
profanityfilter
pyyaml
regex
tensorflow==1.15
tracery
discord-py[voice]
gTTS